Implemented storing the length as more than 1 byte
Changed the Add_Header function to check the length of the value and
save the length as multiple bytes when needed.

The length field is now stored as a vector of bytes.
Saving to file functionality has been updated to reflect this.
Comments were changed to reflect this.

Tests were added to ensure this works correctly.